# Use display cache to boost performance