--- a/src/display.h
+++ b/src/display.h
@@ -9,6 +9,36 @@
 class avatar;
 class Character;
 
+struct disp_overmap_cache {
+    private:
+        tripoint_abs_omt _center;
+        tripoint_abs_omt _mission;
+        std::string _om_wgt_str;
+
+    public:
+        disp_overmap_cache();
+
+        // Returns true if the stored overmap string can be used with the given
+        // center (player) position and mission target.
+        bool is_valid_for( const tripoint_abs_omt &center, const tripoint_abs_omt &mission ) const {
+            return _center == center && _mission == mission;
+        }
+
+        // Rebuild the cache using the validation parameters "center" and "mission"
+        // and store the associated widget string.
+        void rebuild( const tripoint_abs_omt &center, const tripoint_abs_omt &mission,
+                      const std::string &om_wgt_str ) {
+            _center = center;
+            _mission = mission;
+            _om_wgt_str = om_wgt_str;
+        }
+
+        // Retreive the cached widget string
+        const std::string &get_val() const {
+            return _om_wgt_str;
+        }
+};
+
 // The display namespace contains UI string output and colorization functions
 // Some return plain strings or translations, some return a (string, color) pair,
 // and some return a string with colorization tags embedded.